chore(eslint): linting and formatting with eslint and stylistic
eslint deprecated support for formatting because it's too much of a
maintenance hassle to support. They recommend opinionated source code
formatters like Prettier. Well, Prettier's opinion on what is beautiful
code sucks ass so I'm sticking with the crowd that has taken it upon
itself to continue to support formatting rules for eslint.